Crate runwell_entity

Data structures for entity component management.

For efficiency purposes it is not possible to remove once created entities again. Systems are not supported by this ECS system.

Re-exports

pub use self::secondary::ComponentMap;

pub use self::secondary::ComponentVec;

pub use self::secondary::DefaultComponentBitVec;

pub use self::secondary::DefaultComponentMap;

pub use self::secondary::DefaultComponentVec;

Modules

primary
Data structure to create new unique entities.

secondary
Data structures to add, remove and query components for existing entities.

Structs

EntityArena
Primary map to create new entities and store required data for them.

Idx
The index of an entity allocated in an entity arena that holds Ts.

PhantomEntityArena
Primary map to create new entities without the need to directly store data with them.

RawIdx
The raw index of an entity.

Traits

DisplayHook
A hook to customize the Display implementation of type aliases to Idx.