Struct runwell_entity::secondary::default_bitvec::DefaultComponentBitVec

pub struct DefaultComponentBitVec<K> { /* fields omitted */ }

Container where every entity has a bool component that is default initialized to false.

Note

Use this if you need to store bool types as components for your entities efficient and expect most of the entities to have this component.

Prefer using this data structure if the bool components are likely false (default) for most entities.

The API of the DefaultComponentBitVec differs slightly from the DefaultComponentVec because it is not very useful to have bool references which hinders implementing the common Index and IndexMut traits. Use the get and set methods of the DefaultComponentBitVec type instead respectively.

Unbounded iteration over a default component data structure might not be what a user normally wants since they eagerly iterate over the entire key space of their entity which yields approx 2^32 components in total.

Implementations

impl<K> DefaultComponentBitVec<Idx<K>>

pub fn reserve_exact(&mut self, additional: u32)

Reserves the minimum capacity for exactly additional more components to be inserted in the component data structure.

Does nothing if capacity is already sufficient.

pub fn get(&self, key: Idx<K>) -> bool

Returns the current bool value of the keyed entity.

Returns false if the component is not yet set for the entity.

pub fn set(&mut self, key: Idx<K>, new_value: bool)

Sets the bool component of the keyed entity to the new_value.

pub fn replace(&mut self, key: Idx<K>, new_value: bool) -> bool

Replaces the current bool value (if any) of the entity with the new_value.

Returns the old bool value of false if there was no component set before.

pub fn clear(&mut self)

Clears the default component bit-vector for reusing its memory.

pub fn components(&self) -> Components<'_, K>

Returns an iterator yielding the components of the default component bit-vector.

pub fn iter(&self) -> Iter<'_, K>

Returns an iterator yielding the indices and components of the default component bit-vector.

pub fn shrink_to_fit(&mut self)

Shrinks the memory consumption of the default component bit-vector to a minimum.

Note

The operation may costly reallocate heap memory and copy its underlying components.