Module runwell_entity::secondary::map

Container to associate some entities to components in a sparse representation.

Structs

ComponentMap
Sparse secondary map to associated new components for existing entities.

Components
Iterator yielding shared references to the components.

ComponentsMut
Iterator yielding mutable references to the components.

Iter
Iterator yielding keys and a shared reference to their associated components.

IterMut
Iterator yielding keys and an exclusive reference to their associated components.

OccupiedEntry
A view into an occupied entry in a ComponentMap. It is part of the Entry enum.

VacantEntry
A view into a vacant entry in a ComponentMap. It is part of the Entry enum.

Enums

Entry
A view into a single entry in a map, which may either be vacant or occupied.