Struct runwell_entity::secondary::vec::ComponentVec

pub struct ComponentVec<K, V> { /* fields omitted */ }

Dense secondary map to associated new components for existing entities.

Efficiency

Very efficient if the component is very common for the entities. Might be slow for iteration and wastes a lot of space if only a few entities have the component.

Note

The component vector is well suited when a component is very common for entities.

By design all secondary component containers are meant to be easily interchangeable.

Implementations

impl<K, V> ComponentVec<Idx<K>, V>

pub fn reserve_exact(&mut self, additional: u32)

Reserves the minimum capacity for exactly additional more elements to be inserted in the component data structure.

Does nothing if capacity is already sufficient.

pub fn contains_key(&self, key: Idx<K>) -> bool

Returns true if the key is valid for the secondary map.

If the key is invalid the secondary map has to be enlarged to fit the key.

pub fn len(&self) -> usize

Returns the number of components in the secondary map.

pub fn is_empty(&self) -> bool

Returns true if there are no components in the secondary map.

pub fn insert(&mut self, key: Idx<K>, component: V) -> Option<V>

Inserts the component for the key and returns the previous component if any.

pub fn remove(&mut self, key: Idx<K>) -> Option<V>

Removes the component for the key and returns the removed component if any.

pub fn get(&self, key: Idx<K>) -> Option<&V>

Returns a shared reference to the entity at the key.

pub fn get_mut(&mut self, key: Idx<K>) -> Option<&mut V>

Returns an exclusive reference to the entity at the key.

pub fn components(&self) -> Components<'_, V>

Returns an iterator over shared references to the components.

pub fn components_mut(&mut self) -> ComponentsMut<'_, V>

Returns an iterator over mutable references to the components.

pub fn iter(&self) -> Iter<'_, K, V>

Returns an iterator over the keys and shared references to their associated data.

pub fn iter_mut(&mut self) -> IterMut<'_, K, V>

Returns an iterator over the keys and exclusive references to their associated data.

pub fn clear(&mut self)

Clears the component vector for reusing its memory.

pub fn entry(&mut self, key: Idx<K>) -> Entry<'_, K, V>

Gets the given key's corresponding entry in the map for in-place manipulation.

pub fn shrink_to_fit(&mut self)

Shrinks the memory consumption of the component vector to a minimum.

Note

The operation may costly reallocate heap memory and copy its underlying components.