Struct runwell_entity::secondary::vec::OccupiedEntry

pub struct OccupiedEntry<'a, K, V> { /* fields omitted */ }

A view into an occupied entry in a ComponentVec. It is part of the Entry enum.

Implementations

impl<'a, K, V> OccupiedEntry<'a, K, V>

pub fn key(&self) -> Idx<K>

Returns the key from the entry.

pub fn remove_entry(self) -> (Idx<K>, V)

Take the ownership of the key and value from the map.

pub fn get(&self) -> &V

Gets a reference to the value in the entry.

pub fn get_mut(&mut self) -> &mut V

Gets a mutable reference to the value in the entry.

If you need a reference to the OccupiedEntry which may outlive the destruction of the Entry value, see into_mut.

pub fn into_mut(self) -> &'a mut V

Converts the OccupiedEntry into a mutable reference to the value in the entry with a lifetime bound to the map itself.

If you need multiple references to the OccupiedEntry, see get_mut.

pub fn insert(&mut self, component: V) -> V

Sets the value of the entry, and returns the entry's old value.

pub fn remove(self) -> V

Takes the value out of the entry, and returns it.