Struct runwell_entity::EntityArena

pub struct EntityArena<T> { /* fields omitted */ }

Primary map to create new entities and store required data for them.

Note

Use this if all your entities are required to store some additional necessary information or data with them or if the data identifies the entity.

For efficiency and safety reasons it is not possible to remove entities.

Implementations

impl<T> EntityArena<T> where
T: Default,

pub fn alloc_default(&mut self, amount: usize) -> Idx<T>

Allocates the given amount of new entities and initializes them by their default values.

Returns the index to the first entity allocated this way. Indices of allocated entities are guaranteed to be continuous.

Panics

If the operation causes the entity arena to allocate more than or equal to u32::MAX entities in total.

impl<T> EntityArena<T>

pub fn alloc(&mut self, entity: T) -> Idx<T>

Allocates a new entity and returns a unique index to it.

Note

The returned index can be used to query and mutate data of the entity and to add, remove or query associated components of it using secondary data structures.

pub fn clear(&mut self)

Clears the entire entity arena.

Removes all entities. Associated components must no longer be used.

pub fn shrink_to_fit(&mut self)

Reallocates the entity arena to make it take up as little space as possible.

pub fn len(&self) -> usize

Returns the number of created entities.

pub fn is_empty(&self) -> bool

Returns true if no entities have yet been created.

pub fn contains_key(&self, index: Idx<T>) -> bool

Returns true if the entity at the index has been allocated.

pub fn get(&self, index: Idx<T>) -> Option<&T>

Returns a shared reference to the entity at the index if any.

pub fn get_mut(&mut self, index: Idx<T>) -> Option<&mut T>

Returns an exclusive reference to the entity at the index if any.

pub fn indices(&self) -> Indices<'_, T>

Returns an iterator over the indices of the stored entities.

pub fn values(&self) -> Entities<'_, T>

Returns an iterator over shared references to the allocated entities of the entity arena.

pub fn values_mut(&mut self) -> EntitiesMut<'_, T>

Returns an iterator over mutable references to the allocated entities of the entity arena.

pub fn iter(&self) -> Iter<'_, T>

Returns an iterator over the indices and shared references to their associated data.

pub fn iter_mut(&mut self) -> IterMut<'_, T>

Returns an iterator over the indices and shared references to their associated data.