Module runwell_entity::secondary::vec

Container to associate some entities to components in a dense representation.

Structs

ComponentVec
Dense secondary map to associated new components for existing entities.

Components
Iterator yielding shared references to the components.

ComponentsMut
Iterator yielding mutable references to the components.

Iter
Iterator yielding contained keys and shared references to their components.

IterMut
Iterator yielding contained keys and exclusive references to their components.

OccupiedEntry
A view into an occupied entry in a ComponentVec. It is part of the Entry enum.

VacantEntry
A view into a vacant entry in a ComponentVec. It is part of the Entry enum.

Enums

Entry
A view into a single entry in a map, which may either be vacant or occupied.